Title: 앞에서부터 삭제하기 2, Time: 274ms, Memory: 22MB, Status: Passed - CodeTree

+++ 241031/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -21,7 +21,7 @@ public class Main {
         for(int i = n - 2 ; i >= 0 ; i--){
             pq.add(arr[i]);
             sum += arr[i];
-            ans = Math.max(ans, (sum - pq.peek()) / cnt);
+            ans = Math.max(ans, (double)(sum - pq.peek()) / cnt);
             cnt++;
         }
         System.out.printf("%.2f", ans);
